Written simple cache interface (problem)
Problem:
After the a new thread is added, the kernel will clean the entire cache.
If this is replaces with a cache clean by address on the thread control
block and stack, the kernel becomes unstable. Starting an .exe file from
SD card works, but if the .exe file is started twice the kernel krashes.
After reset every program launched from SD card crashes.